capi/src/requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.27
boto3==1.20
